Identify which key in keys table is the signing key.
