+++ services/user/XAdmin/ui/src/hooks/useProducers.ts
@@ -1,12 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getProducers } from "@/lib/graphql/getProducers";
+import { type Producer, getProducers } from "@/lib/graphql/getProducers";
 import { queryKeys } from "@/lib/queryKeys";
 
 import { useConfig } from "./useConfig";
 
 export const useProducers = () => {
-    return useQuery<string[]>({
+    return useQuery<Producer[]>({
         queryKey: queryKeys.producers,
         queryFn: async () => {
             try {
@@ -21,10 +21,21 @@ export const useProducers = () => {
 };
 
 export const useIsProducer = () => {
-    const { data: producers } = useProducers();
+    const { data } = useProducers();
     const { data: config } = useConfig();
+
+    const producers = data?.map((producer) => producer.name);
 
     if (!config || !producers) return undefined;
 
     return producers.includes(config.producer);
+};
+
+export const useMyProducer = () => {
+    const { data } = useProducers();
+    const { data: config } = useConfig();
+
+    if (!config || !data) return undefined;
+
+    return data.find((producer) => producer.name === config.producer);
 };
